Format and rename to value_filters
