+++ imod/common/utilities/value_filters.py
@@ -1,7 +1,6 @@
-import numpy as np
-
-
 from typing import Any
+
+import numpy as np
 
 
 def is_valid(value: Any) -> bool:
