+++ imod/mf6/package.py
@@ -19,6 +19,7 @@ from imod.common.utilities.regrid import (
     _regrid_like,
 )
 from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
+from imod.common.utilities.value_filters import is_valid
 from imod.logging import standard_log_decorator
 from imod.mf6.auxiliary_variables import (
     get_variable_names,
@@ -29,7 +30,6 @@ from imod.mf6.pkgbase import (
     PackageBase,
 )
 from imod.mf6.utilities.mask import mask_package
-from imod.common.utilities.filter_values import is_valid
 from imod.mf6.utilities.schemata import filter_schemata_dict
 from imod.mf6.validation import validation_pkg_error_message
 from imod.mf6.write_context import WriteContext
